+++ src/test/java/unit/market/TestTicker24H.java
@@ -1,11 +1,18 @@
 package unit.market;
 
 import com.binance.connector.client.enums.HttpMethod;
+import com.binance.connector.client.exceptions.BinanceConnectorException;
 import com.binance.connector.client.impl.SpotClientImpl;
+
+import java.util.ArrayList;
 import java.util.LinkedHashMap;
+
+import com.binance.connector.client.utils.UrlBuilder;
 import okhttp3.mockwebserver.Dispatcher;
 import okhttp3.mockwebserver.MockWebServer;
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertThrows;
+
 import org.junit.Before;
 import org.junit.Test;
 import unit.MockData;
@@ -37,9 +44,10 @@ public class TestTicker24H {
 
     @Test
     public void testTicker24HWithSymbol() {
-        String path = "/api/v3/ticker/24hr?symbol=BNBUSDT";
+        String path = "/api/v3/ticker/24hr?symbol=BNBUSDT&type=MINI";
         LinkedHashMap<String, Object> parameters = new LinkedHashMap<>();
         parameters.put("symbol", "BNBUSDT");
+        parameters.put("type", "MINI");
 
         Dispatcher dispatcher = MockWebServerDispatcher.getDispatcher(MockData.PREFIX, path, MockData.MOCK_RESPONSE, HttpMethod.GET, MockData.HTTP_STATUS_OK);
         mockWebServer.setDispatcher(dispatcher);
@@ -47,5 +55,54 @@ public class TestTicker24H {
         SpotClientImpl client = new SpotClientImpl(null, null, baseUrl);
         String result = client.createMarket().ticker24H(parameters);
         assertEquals(MockData.MOCK_RESPONSE, result);
+    }
+
+    @Test
+    public void testTicker24HWithMultipleSymbol() {
+        String path = String.format("/api/v3/ticker/24hr?symbols=%s",
+                UrlBuilder.urlEncode("[\"BNBUSDT\",\"BTCUSDT\"]"));
+        LinkedHashMap<String, Object> parameters = new LinkedHashMap<>();
+        ArrayList<String> symbols = new ArrayList<>();
+        symbols.add("BNBUSDT");
+        symbols.add("BTCUSDT");
+        parameters.put("symbols", symbols);
+
+        Dispatcher dispatcher = MockWebServerDispatcher.getDispatcher(MockData.PREFIX, path, MockData.MOCK_RESPONSE, HttpMethod.GET, MockData.HTTP_STATUS_OK);
+        mockWebServer.setDispatcher(dispatcher);
+
+        SpotClientImpl client = new SpotClientImpl(null, null, baseUrl);
+        String result = client.createMarket().ticker24H(parameters);
+        assertEquals(MockData.MOCK_RESPONSE, result);
+    }
+
+    @Test
+    public void testTicker24HWithInvalidType() {
+        String path = String.format("/api/v3/ticker/24hr?symbols=%s",
+                UrlBuilder.urlEncode("[\"BNBUSDT\",\"BTCUSDT\"]"));
+        LinkedHashMap<String, Object> parameters = new LinkedHashMap<>();
+        String[] symbols = {"BNBUSDT", "BTCUSDT"};
+        parameters.put("symbols", symbols);
+
+        Dispatcher dispatcher = MockWebServerDispatcher.getDispatcher(MockData.PREFIX, path, MockData.MOCK_RESPONSE, HttpMethod.GET, MockData.HTTP_STATUS_OK);
+        mockWebServer.setDispatcher(dispatcher);
+
+        SpotClientImpl client = new SpotClientImpl(null, null, baseUrl);
+        assertThrows(BinanceConnectorException.class, () -> client.createMarket().ticker24H(parameters));
+    }
+
+    @Test
+    public void testTicker24HWithDoubleParameter() {
+        String path = String.format("/api/v3/ticker/24hr?symbols=%s",
+                UrlBuilder.urlEncode("[\"BNBUSDT\",\"BTCUSDT\"]"));
+        LinkedHashMap<String, Object> parameters = new LinkedHashMap<>();
+        String[] symbols = {"BNBUSDT", "BTCUSDT"};
+        parameters.put("symbols", symbols);
+        parameters.put("symbol", "ETHUSDT");
+
+        Dispatcher dispatcher = MockWebServerDispatcher.getDispatcher(MockData.PREFIX, path, MockData.MOCK_RESPONSE, HttpMethod.GET, MockData.HTTP_STATUS_OK);
+        mockWebServer.setDispatcher(dispatcher);
+
+        SpotClientImpl client = new SpotClientImpl(null, null, baseUrl);
+        assertThrows(BinanceConnectorException.class, () -> client.createMarket().ticker24H(parameters));
     }
 }