+++ src/test/java/unit/savings/TestProjectPosition.java
@@ -1,7 +1,6 @@
 package unit.savings;
 
 import com.binance.connector.client.enums.HttpMethod;
-import com.binance.connector.client.exceptions.BinanceConnectorException;
 import com.binance.connector.client.impl.SpotClientImpl;
 import okhttp3.mockwebserver.Dispatcher;
 import okhttp3.mockwebserver.MockWebServer;
@@ -13,7 +12,6 @@ import unit.MockWebServerDispatcher;
 import java.util.LinkedHashMap;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertThrows;
 
 public class TestProjectPosition {
     private MockWebServer mockWebServer;
@@ -23,18 +21,6 @@ public class TestProjectPosition {
     public void init() {
         this.mockWebServer = new MockWebServer();
         this.baseUrl = mockWebServer.url(MockData.PREFIX).toString();
-    }
-
-    @Test
-    public void testProjectPositionWithoutParameters() {
-        String path = "/sapi/v1/lending/project/position/list";
-        LinkedHashMap<String, Object> parameters = new LinkedHashMap<>();
-
-        Dispatcher dispatcher = MockWebServerDispatcher.getDispatcher(MockData.PREFIX, path, MockData.MOCK_RESPONSE, HttpMethod.GET, MockData.HTTP_STATUS_OK);
-        mockWebServer.setDispatcher(dispatcher);
-
-        SpotClientImpl client = new SpotClientImpl(MockData.API_KEY, MockData.SECRET_KEY, baseUrl);
-        assertThrows(BinanceConnectorException.class, () -> client.createSavings().projectPosition(parameters));
     }
 
     @Test